fix: replace config limit with sdk.batch.size
